+++ ui/packages/shared/components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.108](https://github.com/parca-dev/parca/compare/@parca/components@0.16.107...@parca/components@0.16.108) (2023-03-07)
+
+**Note:** Version bump only for package @parca/components
+
+## 0.16.107 (2023-03-07)
+
+**Note:** Version bump only for package @parca/components
+
 ## 0.16.106 (2023-03-02)
 
 # 0.16.0 (2023-03-01)