+++ ui/packages/shared/profile/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.133](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.132...@parca/profile@0.16.133) (2023-03-07)
+
+**Note:** Version bump only for package @parca/profile
+
+## 0.16.132 (2023-03-07)
+
+**Note:** Version bump only for package @parca/profile
+
 ## [0.16.131](https://github.com/parca-dev/parca/compare/@parca/profile@0.16.130...@parca/profile@0.16.131) (2023-03-06)
 
 **Note:** Version bump only for package @parca/profile